[AIP-21] Move contrib.sensors.redis_key_sensor providers.redis.sensors.redis_key
