+++ airflow/providers/redis/sensors/redis_key.py
@@ -16,14 +16,24 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.redis.sensors.redis_key`."""
+from airflow.providers.redis.hooks.redis import RedisHook
+from airflow.sensors.base_sensor_operator import BaseSensorOperator
+from airflow.utils.decorators import apply_defaults
 
-import warnings
 
-# pylint: disable=unused-import
-from airflow.providers.redis.sensors.redis_key import RedisKeySensor  # noqa
+class RedisKeySensor(BaseSensorOperator):
+    """
+    Checks for the existence of a key in a Redis
+    """
+    template_fields = ('key',)
+    ui_color = '#f0eee4'
 
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.redis.sensors.redis_key`.",
-    DeprecationWarning, stacklevel=2
-)
+    @apply_defaults
+    def __init__(self, key, redis_conn_id, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.redis_conn_id = redis_conn_id
+        self.key = key
+
+    def poke(self, context):
+        self.log.info('Sensor checks for existence of key: %s', self.key)
+        return RedisHook(self.redis_conn_id).get_conn().exists(self.key)
