+++ scripts/ci/pylint_todo.txt
@@ -88,7 +88,7 @@
 ./airflow/providers/mongo/sensors/mongo.py
 ./airflow/sensors/python.py
 ./airflow/contrib/sensors/qubole_sensor.py
-./airflow/contrib/sensors/redis_key_sensor.py
+./airflow/providers/redis/sensors/redis_key.py
 ./airflow/contrib/sensors/redis_pub_sub_sensor.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_base.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_endpoint.py
