+++ tests/providers/redis/sensors/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.redis.sensors.redis_key`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.redis.sensors.redis_key import RedisKeySensor  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.redis.sensors.redis_key`.",
-    DeprecationWarning, stacklevel=2
-)
