+++ tests/providers/redis/sensors/test_redis_key.py
@@ -23,8 +23,8 @@ import unittest
 import pytest
 
 from airflow import DAG
-from airflow.contrib.sensors.redis_key_sensor import RedisKeySensor
 from airflow.providers.redis.hooks.redis import RedisHook
+from airflow.providers.redis.sensors.redis_key import RedisKeySensor
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
